requirements.txt
scipy~=1.15.2
numpy
pandas~=2.2.3
pydicom~=3.0.1
matplotlib~=3.10.0
natsort~=8.4.0
scikit-learn~=1.6.1
pillow~=11.0.0
nibabel~=5.3.2
torchio
umap-learn
pytorch-fid
imageio